Change overplotFiberTrace.py to work with new Flat and Arc
The previous Flat and Arc (visit numbers 4 and 5) in drp_stella_data
have been replaced with the new dithered Flats and a new Arc with
the same FiberTraces illuminated as in the dithered Flats. This
meant that overplotFiberTrace.py needed some changes to work with
these new fits files. I also added a paragraph to overplot the
FiberTrace in the Arc.