CurveFitting::PolyFit: bug: same yFit array for both layers
So far the keyword parameters were simply passed from the
top-layer PolyFit with sigma rejection to the bottom layer PolyFit
what caused a problem when data actually got rejected. This
was because the size of the yFit array would be larger than the
size of the data arrays. Fixed by using 2 different yFit arrays
for both PolyFit layers.