Change FiberTrace coefficients from double to float
As the tests have shown, a FiberTrace reconstructed from
the persisted FiberTrace showed an offset in one row by one
pixel, caused by persisting the double FiberTrace function
coefficients as floats. changing the coefficients from double
to float seems to solve the problem. If not, we'll need to change
back to doubles and persist doubles in pfsFiberTrace.