Math::minIndex: Use std functions and add documentation
The code changes should make the function faster, easier to
read, and less error prone. I changed the function originally
in order to find a bug which turned out to be somewhere else,
but I think the code changes are worth keeping. We should also
consider to change related functions accordingly.