Controls::DispCorControl: Add parameter maxDistance
Emission lines whose measured center positions in pixels is more than
this parameter value away from the predicted positions will be rejected
from the calibration procedure. The main difference to 'searchRadius'
is that searchRadius is an int while maxDistance is a float.
While searchRadius is used to determine the range to fit, maxDistance
is then used to check the result of the fit.
Spectra::Identify: use DispCorControl.maxDistance instead of
previously hard coded number.