fix FiberTraces::assignITrace and FiberTraces::findITrace
findITrace used to look for a FiberTrace in redFiberTraces.fits
within a certain limit of its x- and yCenter and could fail if
there was none within that limit. This has changed to return the
closest FiberTrace which will always return a number. findITrace
has therefore been changed to return a size_t instead of int
(used to return -1 if it failed). assignITrace was changed to
return void instead of bool.
From both functions the parameter yCenter has been removed as
it was not used for anything (anymore).