FiberTrace::extractFromProfile: include background in spectrum
So far we constructed a Spectrum for the object and a Spectrum
for the background (sky if slit), but only returned the object
Spectrum. We have a place holder for the background spectrum
in the Spectrum class, which we now use for the background.
However, we still don't have a placeholder for the uncertainties
in the background.