CurveFitting::PolyFit: Fix bug in MeasureErrors
There was a bug that the measure errors for values
rejected by the PolyFit sigma rejection algorithm
were not removed from the list of measure errors,
only the rejected values themselves got removed from
the iteration procedure. This is fixed now by creating
a new list of keyword values for the iteration procedure
and removing the measure errors of the rejected values
from the iteration.