Add tests for PolyFit
Include test for basic functionality of PolyFit in Utils.cc:
As the vector for the keyword values in PolyFit (vector of void
pointers) couldn't be swigged I moved the tests to Utils.cc. These
tests in Utils.cc are being called by tests/Math.py

Include test for PolyFit in Spectra.py:
The test checks if after manually putting a cosmic into the wing of
an emission line the line correctly gets rejected from the fit and
the appropriate mask flag is set.

Math.py: move setUp to testWhere
and delete setUp and tearDown. The 1 array declared in setUp was
not needed by testPolyFit. I therefore moved its declaration into
testWhere and deleted setUp and tearDown.