Use lsst::afw::image::MaskType instead of hard-coding mask image type.
This is important as we just changed from 16- to 32-bit ints to the old
"unsigned short" was incorrect.

Note that some of the "unsigned short"s in the codebase are just
boilerplate template instantiations that are probably unused, but
because "int" is also instantiated, if we changed to use MaskType
we'd get a double instantiation which won't compile.

Also change MaskU to Mask as the former no longer exists.