We don't need our own __call__, thereby sidestepping the problem.
As proposed by Paul Price.  Unfortunately cmdLineTask cannot handle
writing the resulting metadata (task.writeMetadata is passed a list
of dataRefs);  as a workaround we disable writing metadata.  That's
OK for now as there's nothing useful in the metadata -- but it's
something that we will need to solve properly (??superTask??)