Quite a lot of cleanup.
Spectra and FiberTraces are now all the same height (as was always assumed in the data model for pfsArm) (so no need for
{min,max}Y)

Renamed "size()" to "getNpix()" and "getNtrace()"

Fixed setting the variance and mask.

Stopped getAllXXX from returning the transpose of what we want
(which also required fixing the declaration of _covar)

Cleaned up the code for allocating ndarray slices.

Removed the odd extra allocation and copy.

Remove unused ctors for SpectrumSet; No need to make copies, so
need for the cctor.