FiberTrace: remove private calculation methods
The private calculation methods in FiberTrace existed to construct the
FiberTrace model, which is much simpler than the list of private members
implies. All we need for a FiberTrace is the trace profile and a fiberId.
Everything else was inserted to simplify the API of the various methods
called in the process of determining the model.
Now all of that has been moved out to findAndTraceApertures, which is
the only place it was used. Some vestigial parts (vectors of vectors of
points) have been stripped out.
The result is a much cleaner FiberTrace class, which could now be
constructed using a different algorithm or implementation than the
one that was hard-wired in. The state of the object is now much
clearer at all times, and simpler to represent.