fitProfile2d: mask bad results
We've long known we need to mask the spectral element when the fit
returns a negative chi^2; finally did it. Reordered the elements in
the tuple coming back from fitProfile2d to correspond to the classic
image,mask,variance order from lsst::afw::image::MaskedImage.
Bad spectral elements are masked as BAD; we probably want this to be
configurable, but this will do for now.