refactor flat field normalisation
Previously, each fiber was independently normalised, which means
that flux calibrating one doesn't allow you to flux calibrate
anything else (well, we have a throughput measurement in the
DetectorMap, but that's not a function of wavelength). Instead,
normalise all fibers to a continuum fit of the mean spectrum of
the fibers.
This keeps the fibers tied together.