Add SpectrumSet.plot method
This involves a bit of refactoring of Spectrum.plot so they can share
code. Renamed the plotBackground and plotReferenceLines kwargs to
doBackground and doReferenceLines.

SpectrumSet.plot doesn't have all the features that Spectrum.plot does,
as I decided that plotting the background and reference lines would just
make the plots too confusing.