identifyLines: add interactive identification
If the wavelength solution in the DetectorMap is significantly
inaccurate, the automated identification of lines (Spectrum.identify)
may fail or produce non-optimal results. This module allows an
interactive (i.e., manual) identification of lines that will be fit
and used to modify the DetectorMap, allowing the subsequent automated
line identification and wavelength calibration to succeed.