improve splines in DetectorMap
We had a constant number of knots for both the xCenter and wavelength
splines, and when presented with a lookup table it would take every
Nth pair as the knots, rather than doing a fit.
Instead, take the entire lookup table as the knots. Then we just use
the spline to do intra-pixel interpolation, and we preserve the full
fidelity of whatever fit has been performed when we construct our
spline.
This requires a different persistence format, so existing DetectorMaps
on disk are invalidated.