add PairMatchIdentifyTask
The current algorithm used in IdentifyTask (and implemented in
Spectrum.identify) simply centroids lines where they are expected to be.
This is not robust against modest shifts in wavelength, or slightly
inaccurate wavelength solutions; but we get such cases during construction
of the instrument.
This algorithm matches pairs of lines, and hence is more robust
against inaccurate wavelength solutions. The centroid fitting
uses astropy, which means we don't have to maintain the fitter;
and all the identification is done in python, which is much
easier to debug than the current C++ version.