calibrateWavelengths: fix plots
The plot of linear residuals did not include offsets of the points from
the fit, which was very misleading.
The plot of wavelength as a function of pixel included the linear residuals
twice (onec because it's already in 'spec.wavelength', and then another
included explicitly), which again was misleading as the line didn't go
near the points sometimes.