add definition of Lsf
Lsf is an abstract base class based on LSST's Psf class, but
brought down to one dimension. It presents the same interface
as a Psf (duck typing) so that it can be used in the place of
a Psf (in python only), with additional methods specialised
for 1D use.

This includes a Kernel1D class (for convolving by your Lsf, or
inserting the Lsf at a particular point in the spectrum) and
a couple of basic concrete Lsf classes for demonstration.