findAndTraceApertures: make width limits configurable
findAndTraceApertures fits the fiber profile with a Gaussian, and sets
acceptance limits on the parameters. These limits were hard-coded to
1/4*apertureFwhm < sigma < apertureFwhm, and offset < 1. Made these
limits configurable through the control object, and set looser defaults
to allow finding traces with a wider range of widths.

These looser default limits might be ratcheted down once the instrument
is stable.