identifyLines: modify algorithm to work with missing lines
The current algorithm uses Spectrum::identify to fit lines on
the spectrum at the expected positions of reference lines.
However, if there is no actual line in the spectrum at the
expected position of the reference line (e.g., the line is
fainter than the detection limit), this centroids on noise
peaks, compromising the wavelength solution. (It is possible
to put a threshold on the 'guessedIntensity' of the reference
lines, but this makes a configuration parameter a function of
the data, e.g., exposure time, lamp setup, etc.)

This commit introduces a new algorithm for identifying lines
on the spectrum: find lines in the spectrum, and match those
against the reference lines (similar to what we do for
astrometric solutions with LSST/HSC). This uses the new Task
for finding lines that we introduced for bootstrapping the
detectorMap (the threshold is configurable, and respects
the noise). Then we match with the nearest line within
some limit (default 0.3nm).

The old algorithm is preserved within the OldIdentifyLinesTask.