findLines: optionally subtract continuum before finding lines
The HgAr arc has a strong continuum, and the algorithm here was just
checking that the tip of the peak (including any continuum!) was above
threshold, not that the height of the peak (above the continuum) was
above threshold. This meant we picked up noise spikes.

Modified the API for FindLinesTask.run to return a Struct including the
continuum (because someone in the future may want it, at which point an
API change will be much more disruptive; and it's the right thing to do).