identifyLines: add exclusion radius around reference lines
To avoid confusing reference lines, optionally exclude any reference lines
that are within an exclusion radius of each other.

This feature is currently disabled, as I don't believe it's necessary for
getting a good wavelength solution (the clipping in the wavelength fit
is usually sufficient to get rid of any misidentifications if we're not
matching against noise), but I implemented it and it may prove useful in
the future.