incorporate FluxTable into pfsObject and pfsSingle
This requires some changes to the plumbing, because we want to build
the FluxTable from unresampled arm spectra, but we were previously
building the coadds from the merged-arm spectra for convenience.

Made FocalPlaneFunction care about wavelength (like SubtractSky1DSolution,
so we no longer need that) so we can use it to correct both resampled and
unresampled spectra. Added methods to MeasureFluxCalibrationTask and
SubtractSky1DTask (which use FocalPlaneFunction) to apply their corrections
to either PfsSpectra (e.g., a pfsArm) or PfsSpectrum (e.g., pfsSingle).
Changed CoaddSpectraTask so that it applies the flux calibration and
1D sky subtraction to individual arm spectra instead of doing the merge.