WavelengthSampling: use WavelengthArray
This provides a modest savings in space by allowing us to write
the wavelengths as a WCS rather than the full array. Changed
the parameters so that we specify a length rather than the
wavelength spacing, consistent with how a WavelengthArray is
specified. Set the length to give us 0.8 A/pix as proposed in
datamodel.txt