FiberTraceSet: add method to extract spectra
After combining traces, neighbouring traces may overlap, so we
need to solve the least-squares problem for each row in the
image. The matrix in this least-squares problem is symmetric
and tridiagonal, which can be solved faster than a regular
matrix inversion. We also get the matrix inverse, which allows
us to set the covariance elements.