refactor FindLinesTask to also measure line flux,width
FindLinesTask was introduced to measure emission line centroids for
wavelength calibration, but there's more information available in the fit
parameters that users might want. Refactored it so that each method does
one thing, and changed the return values so that all the fit parameters
are returned. Added a 'runCentroids' method that preserves the original
interface of the 'run' method. Also added tests.