bootstrap: fit/update CCDs separately
The detector for b, r, m arms is actually two CCDs, and the
separation and rotation may be slightly different than for
the simulator. By fitting the two sides separately, we should
get a better solution.

The alternative would be to fit the separation and rotation
explicitly. That should be more accurate, but fitting them
separately is easier.