NevenPsf: allow configuration from MeasurePsfTask
The parameters for NevenPsf.build were provided in the API, but I want to
allow MeasurePsfTask to provide None for each, as otherwise requires
storing the default values more than once (once in the API and once in
MeasurePsfTask). Moved default values into the function body and added a
directory option for full control. MeasurePsfTask can now set these
parameters through the config.