constructFiberTrace: check fiberIds match pfsConfig and fix if desired
If the fiber identifications are wrong, we're going to get
in trouble in downstream processing. Check the fiberIds
that come from matching with the DetectorMap (in
findAndTraceApertures) with those in the pfsConfig. This
is a simple algorithm, and may need to be upgraded when
we start hiding fibers behind dark spots, if we manage
to find and trace those.