subtractSky2d: reduce verbosity
Warning messages about lines that can't be measured or
subtracted are fairly common (there are lots of lines
off the image, or just off some fibers); demote these
to 'debug' level.

The huge string of measurements and the long string of
the model are also more appropriate for 'debug' level.