findTracePeaks: operate on flux rather than S/N
The images we're finding traces on can have a variety of
rubbish, like scattered light or crosstalk, that we don't
want to detect. This is usually at a low level, so it makes
sense to set a threshold in terms of flux rather than
signal-to-noise.

If you want the function to threshold by signal-to-noise,
then give it a signal-to-noise image.