buildFiberTraces: make trace association more robust
Because we allowed multiple peaks to match to a single trace, it was
possible for a cosmic ray at low y to link multiple fibers as a single
trace before the fibers had any peaks themselves. This meant that fibers
would be dropped because they were joined together.

Instead, this introduces a scheme that allows only a single peak from each
row per trace. If there are multiple peaks in a row that want to be
assigned to the same trace, we take the one with the most matches in the
last 'associationDepth' rows.