reduceArc: support xCenter tweaking from arc
The fiberTrace and the arc may have slightly different x centers (e.g.,
due to hexapod moves), and if we want to work on the arc image (e.g., for
sky subtraction investigations) then we need to have the updated centers
apply to the arcs rather than the fiberTrace.

Added a mechanism to tweak the x centers using the arcs, and made this the
default.