fitLine: add support for errors
We have been assuming that all centroids have equal errors, but brighter
lines obviously have smaller errors, and we should respect this in our
wavelength calibrations. The first step is to get errors out of the fitter,
which is what we do here. We don't put variances into the fitter, since
the variances include Poisson noise that varies with the flux of the line,
and would hence introduces biases as a function of flux. But the errors
that come out can be used to weight the wavelength calibration.