introduce GlobalDetectorMap+SplinedDetectorMap
These are subclasses of BaseDetectorMap. This allows the creation of
further subclasses (e.g., DifferentialDetectorMap) if we consider that
necessary.

SplinedDetectorMap is a direct descendent of DetectorMap, where each fiber
has xCenter and wavelength as a function of row in a spline.

GlobalDetectorMap (using the new GlobalDetectorModel class) models the x,y
position as a function of fiberId,wavelength (using a 2D Chebyshev
distortion field, and individual spatial,spectral offsets for each fiber).