constructSpectralCalibs: fix DATE-OBS in output header
Now we are including the times along with the dates, the LSST
CalibTask.calculateOutputHeaderFromRaws produces a DATE-OBS with two
times (the proper one first, and then 00:00:00.00 because it thinks the
date doesn't include a time). We monkey-patch the method, and strip the
extra time. This suppresses warnings from parsing DATE-OBS downstream.