CentroidLinesTask: find peak near expected position
SdssCentroid expects that the expected position is already near the peak.
If the detectorMap is off by a couple of pixels, it could fail to find a
good centroid. So we look for a peak in the convolved image near the
predicted position, and use that as the starting point for the proper
centroiding.