FitFocalPlaneTask: extend to respect masks and variances
FitFocalPlaneTask has been a placeholder, and its shortcomings have started
to be felt in the form of false emission lines from a single bad pixel in
a sky fiber. This commit upgrades its capabilities, so that it implements
rejection and handles variances (input and output): essentially, it now
does a weighted coadd.

This invalidates any existing sky1d and fluxCal products on disk, as the
I/O has changed to support the mask and variance. The API has also changed,
so updated SubtractSky1dTask and MeasureFluxCalibrationTask to use the new
capabilities.

For all of this, the implementation is still somewhat of a placeholder,
since it doesn't do any fitting across the focal plane.