subtractSky1d,measureFluxCalibration: add systematics term to variance
The spectra that we're going to fit for the sky1d or fluxCal have been
resampled, which means that some variance has been lost as covariance.
To attempt to account for this and any other systematic problems, add a
fraction of the flux value to the variance. This helps avoid excess
rejection in the FitFocalPlaneTask.