refactor GlobalDetectorMap for faster fitting
GlobalDetectorModel (which provides the algorithmic backing for
GlobalDetectorMap) fit a model to the data using Minuit2. This was slow
when dealing with lots of inputs (e.g., a couple of exposures with a full
fiber density) because of the fitting algorithm chosen (non-linear fitting
with gradient descent) and the need to perform multiple fitting iterations
in order to avoid parameter degeneracies. This approach worked (at least,
for inputs on the scale of a few exposures of ~10 fibers each) but did not
perform at scale.

Refactored the fitting to use a linear fitting algorithm (matrix inversion)
which is much faster, especially for large input datasets. The data
contained within the model has changed (different parameters), and so this
invalidates any GlobalDetectorMap files on disk (but GlobalDetectorMap
hasn't been around long, so this should not cause much trouble).