DetectorMap: add vectorized findPoint,findWavelength
Iterating in C++ is much faster than in python, even when there's a virtual
function call involved. However, pybind11 doesn't like us having a virtual
function that takes scalars and a non-virtual function that takes vectors
and iteratively calls the scalar version (because the virtual function has
to be re-defined in the subclass, which then doesn't look to the base
class for overloads). Instead, use a different name for the virtual
function, and define all the functions in the base class. This way, there's
no overloads to confuse pybind: everything's handled in C++.