DetectorMap: refactor to use I/O from datamodel
By using the I/O in datamodel, we ensure that we're always writing the
correct format, but it means a bit of a refactor. We no longer have to
determine the subclass of data on disk (that's done in the datamodel I/O
code); now we just have to convert to/from the datamodel representation.