buildFiberProfiles: extend scoring system in association
In associating peaks with traces, peaks are assigned as score
which is the horizontal distance between the peak and the
last position of the trace. It's possible for those both to
be the same (even both zero, when CRs and masking are involved)
and so we need a secondary means of scoring to break ties.
We choose the negative number of appearances of that trace in
the association memory (negative, so a higher value becomes
a better score), so a thin cosmic ray shouldn't interfere
much.