SplinedDetectorMap: optimize findPoint
The SplinedDetectorMap::findPoint method, which is used a lot by
DifferentialDetectorMap, was slow because it calculates the wavelength
for all rows of the fiber.

Instead, we introduce an additional spline in the implementation, which is
the inverse of the wavelength as a function of row.

This change  allows us to reduce threshold on the test a little, so it
must be doing even better than before.