Spline: always use double for coefficients
Regardless of what's being passed in and out, we want everything
backed by double-precision math.

Don't think this commit is entirely done. I did this while
worrying about something else and then it wasn't important
for that, but I'm putting it here in case it saves me doing
the same work again later.