extractSpectra: include all fiberIds specified
It's annoying to have a different order or set of fiberIds between the
pfsConfig and the extracted spectra, so we now allow specification of
fiberIds to include in the returned spectra (any missing are filled with
NANs).

Of course, for spectrographs other than SM1, there's no way for the indices
of fiberId arrays in pfsConfig and the extracted spectra to match up, but
SM1 is the main instrument we're working with right now.