DifferentialDetectorMap: fix measureSlitOffsets
Identified two problems:
1. Slit offsets weren't being updated in the DifferentialDetectorMap
(they're measured using the base SplinedDetectorMap, but the values
didn't find their way into the DifferentialDetectorMap that was
holding it). Not sure whether this caused problems with the
positions, but it made it appear that there were no offsets
because the offsets logged by MeasureSlitOffsetsTask come from the
DifferentialDetectorMap.
2. Measurement of the slit offsets was not respecting the
distortion model that is used in addition to the base
SplinedDetectorMap, so slit offsets were attempting to soak up
power from the distortion model, and then we were applying the
distortion model on top of those slit offsets when determining
positions; this results in bad positions.