traces: don't assume that the last valid pixel was x-1
In the presence of a block of masked pixels, the last valid pixel was
several pixels ago, because we skip over masked pixels completely; but the
code assumed it was the last pixel (x-1). This meant that when a block of
masked pixels interrupts a trace that has already been identified, the
peak was at the end of the masked pixels, and not at the position of the
maximum unmasked flux.

This caused problems in the association of peaks when there was a block
of masked pixels between the traces, because the peak positions could be
on the trace to the right.

Fixed by keeping track of the position of the last valid pixel.