fiberProfile: use double for calculations
The one place where we weren't using double-precision floating point for
calculations was in the image (which was single-precision since we were
going to return it). Any operations on the image used single-precision,
which might be detrimental.

Changed this to use a double-precision image, and convert to
single-precision only at the last moment.