constructFiberProfiles: trust detectorMap to make fiber traces
There are subtle differences between the polynomial fits to
the trace centroids and what the detectorMap predicts. These
might be due to systematic noise in the centroids or
insufficient model flexibility; in either case, the
detectorMap is the correct representation. If the
detectorMap is slightly off, then we can update the
detectorMap through iteration.