remove old, unused C++ functions from a previous era
In particular, this involves removing Spectrum::identify, replacing
FiberTrace.extractSpectrum with an alternate implementation, and removing
findAndTraceApertures, along with various supporting functions. Much of
this has been replaced with alternate implementation (generally in python)
that we have been successfully using for a while.

This was precipitated by compilation errors with clang 12.0.0 (on Macs) in
some of the code that's been removed. We don't want to bother supporting
code that we're not using.