hoist ReferenceLine to python as ReferenceLineSet
The ReferenceLineSet is implemented the same way as ArcLineSet. The new
class doesn't confuse reference properties (e.g., wavelength) with
measured properties (e.g., position in the spectrum) and is not closely
tied to the Spectrum class. It also uses a bitmask for the 'status'
member, which has some pre-defined values in the new ReferenceLineStatus
enum. The class can read/write line lists in text format, and it doesn't
filter lines that are flagged, so the user gets all lines and can select
them as they think suitable.

This involves:
* stripping the referenceLines member out of the Spectrum class.
* modifying various functions to operate with reference lines explicitly,
  rather than expecting them to be contained within the Spectrum class.
* modifying IdentifyLinesTask to produce ArcLineSet (i.e., measured lines)
  rather than ReferenceLine (i.e., line list).
* moving the line list reading code into ReferenceLineSet (the main I/O)
  and ReadLineListTask (filtering).
* removing the CalibrateWavelengthsTask, as removing the unused class is
  easier than updating it.
* add configuration options to FitDifferentialDetectorMapTask and
  MeasureSlitOffsetsTask to filter by the line status bitmask.
* reorganise ReduceArcTask to remove use of IdentifyLineTask: it's not
  needed any more, and there's no reason to maintain it.