add DistortedDetectorMap
This is a modified version of the DifferentialDetectorMap, except instead
of defining the distortion field in terms of fiberId,wavelength (and
therefore requiring scaling to xi,eta space, which was confusing), the
distortion field is defined in terms of x,y (of the base detectorMap).
This is much simpler, and more natural (it's what RHL wanted me to do a
long time ago, but I was influenced by the GlobalDetectorMap idea).

Decided to add this new Distorted DetectorMap class instead of upgrading
the DifferentialDetectorMap class to avoid invalidating existing data
products. DifferentialDetectorMap is now deprecated, and we'll remove it
eventually.

Much of the code supporting DifferentialDetectorMap has been copied.
The new DetectorDistortion class replaces the GlobalDetectorModel class.