add AdjustDetectorMapTask
This fits a low-order correction to the detectorMap, using the position
of the traces and lines on the image. It is intended to be used to correct
for small changes to the detectorMap, using the science images themselves.

Because it's using the traces as well as emission lines, it uses a
non-linear fitter that's a bit slower than the linear fitter used when
we fit the full detectorMap, but that works out well because we only need
to correct the lowest orders.

This replaces the use of MeasureSlitOffsetsTask in reduceExposure,
which was attempting to correct for the same problem.