restructure NevenPsf to interpolate over individual fibers
NevenPsf has been ignoring the fiberId values, and using the x,y values in
the interpolation. Here we introduce a new interpolation algorithm that
operates per fiber; this requires some restructuring for efficient fiber
lookup.

We store a (sorted) array of wavelength, y and PSF image for each fiber.
Given a fiberId,wavelength (well, not really: PSFs operate in x,y, so we
have to convert to fiberId,wavelength; this may be inefficient, but we'll
worry about it if/when it is determined to be so), we can interpolate in
y for that fiber.

We're also switching to Neven's latest data version, Jan0821_v3. The
detectorMap required for the test data is much smaller than the previous
one, so we don't need the caching any more (in fact, the caching introduces
memory leaks). Somehow, we managed to improve the interpolation precision
from 1e-8 to 1e-9!