record extracted spectra in counts
We have been normalising by the quartz spectra (provided in the
fiberProfile), but we want the pfsArm files to have flux in counts, and
pfsMerged to have flux in something close to counts. In order to avoid
the risk of using the wrong normalisation later in the pipeline, we now
carry the normalisation around in the pfsArm (and pfsMerged, and therefore
the Spectrum and PfsFiberArraySet classes).

I've tried to fix all the places where we want to use the normalised
flux instead of the regular flux (e.g., when finding lines, using the
normalised flux removes the grating response) and where we manipulate the
flux and variance, so that the fluxes in the output products are accurate.