MergeArmsTask: use normalisation array instead of single value
By normalising to a single value, the flux in the pfsMerged
of a quartz is flat, which doesn't look at all like what we
were aiming for: a close-ish representation of the pfsArm
fluxes.  Instead, make the normalisation be an array, and
make it look as much like the quartz as we can (except
smoothed).