rewrite adjustDetectorMap to use linear solver
adjustDetectorMap used a non-linear solver to fit both line and trace
centroids. This worked nicely, but was slow because of the non-linear
optimisation.

Rewrote the task completely from scratch to use a linear solver (in fact,
we now subclass the FitDistortedDetectorMapTask and use the same fitting
process). We use lines for fitting if they are available, and use trace
centroids (masquerading as lines) only if there are no lines (because
lines are so much more accurate and they allow us to solve for wavelength
shifts too).

This reduces the adjustDetectorMap runtime from 48 sec to 19 sec (3x
faster).