add C++ function to centroid image with SdssCentroid
This will be used to centroid the oversampled PSF images using the same
centroiding algorithm that we use on the arc lines, so the PSF and the
DetectorMap should have a consistent definition of where the center is.

I've added multiple overloads with python bindings, so they can be used
in development of PSF models.

This adds meas_base as a formal dependency, since that's where SdssCentroid
is implemented.