revise ReferenceLine.plot
ReferenceLine.plot was copied from the old plotReferenceLines function,
but it wasn't really updated to work well with the new ReferenceLine
class. For example, the ReferenceLine class doesn't have "fitPosition"
or "guessedPosition" values, so there's no point in making what's being
plotted configurable: it's always the actual wavelength of the line.
Also, the status enums weren't being accessed correctly.

We do still want to be able to plot against both pixels and wavelength,
so added support for that.