refactor FocalPlaneFunction, FitFocalPlaneTask
We're getting more serious about FocalPlaneFunction, and different styles
of fitting, as we're going to want to gradually add more functions with
increasing functionality.

Moved FocalPlaneFunction into its own file, and refactored to be more
simple to write subclasses. Added rejection and iteration to
FitFocalPlaneTask so users don't have to implement this common feature
themselves.