refactor sky subtraction, flux calibration, mergeArms, coaddSpectra
Sky subtraction and flux calibration are users of FitFocalPlaneTask,
which has recently been refactored. Much of the SubtractSky1dTask and
MeasureFluxCalibrationTask functionality has mostly moved into the
FitFocalPlaneTask (iteration and rejection), so they can be replaced by
thin functions to apply the results, and a small bit of code above them
that will

This requires a bit of refactoring of mergeArms and coaddSpectra, since
these are the main users of the results of sky1d and fluxCal. In
particular, we no longer allow 1d sky subtraction to be performed on
resampled images, since we have oversampled splines; this simplifies the
code because coaddSpectra no longer has to know when to apply the 1d sky
subtraction.