ReferenceLineSet: make exclusion zone absolute
No longer allow the brightest line within the exclusion zone, but remove
any and all lines that are within the exclusion zone. Whether the lines
really are blended or not is a function of the intensity and resolution
and distance apart and the signal-to-noise; this is too complicated for
this function, so remove everything in the interests of purity.