centroidLines: add deblending
Lines are measured individually, but we were doing so on the image with
all lines present. This may affect the centers that we measure, because
a nearby line can get flux into the line of interest when we convolve.

The classic solution to this (from LSST, SDSS) is to use a deblender that
replaces nearby sources with noise. Our deblender can be simple, since all
sources are on the fiber traces. Deblending in the spatial dimension
is trivial, so we only need to deblend in the spectral dimension, which we
do simply by looking for the saddle point when moving away from the line
of interest.