add DoubleDetectorMap
This is a DistortionBasedDetectorMap that uses two separate distortion
fields for the two CCDs. It doesn't seem to make a great difference, but
Jim Gunn says this is how we should treat the distortion, since the CCDs
have different hills and valleys, so the distortion field is not continuous
across the center line.