ArcLineSet: add asPfsFiberArraySet
Representing line measurements as a set of spectra will help us when we
want to fit models of line intensities, since sets of spectra are the
input format for the FitCoalPlaneTask. This will have immediate
application to aperture corrections, but I expect will also be used for
sky2d later.