add aperture corrections to line photometry
If the fidelity of the PSF is wavelength-dependent even within a single
fiber, then we need to apply aperture corrections. We measure the aperture
corrections directly from the data (assuming there are isolated lines),
and then fit a polynomial as a function of wavelength to the aperture
corrections for each fiber.

We persist the aperture correction in case it's needed to understand things
later. The aperture correction applied to the 'lines' has to be undone in
subtractSky2d because we are inserting PSF fluxes.

I'm a little concerned I had to loosen the tolerance in the unit test after
applying the aperture correction, and that I don't understand why this was
necessary.