PfsFiberArray/Set: use temporaries in imul,itruediv
If we do pfsArm /= pfsArm.norm, we change pfsArm.norm in the
process of performing the operation, which means that the flux
and the variance are divided by different values.

To work around this, and any similar situation, copy the array
first, to ensure we don't share memory with any element of
the object we're changing.

In the process, changed implementation of division to be the
inverse of multiplication, to reduce code duplication.