Selecting only science fibers from pfsConfig.
This is done both in constructFiberProfilesTask and
reduceExposureTask. This is a TEMPORARY  workaround
to address recent problems with processing PFI data,
where the pfsConfig now includes engineering fibers.

This problem should be addressed once detectormaps applicable
for PFI data, that include engineering fibers, are generated,
and if necessary modifications to DRP such that it is robust
against non-illuminated or broken fibers is in place.