fitDistortedDetectorMap: increase default soften
The 'soften' parameter serves to prevent points from being
rejected due to underestimated errors. We've been measuring
softening values after all the fits that are typically around
0.03 pixels, so we should increase the rejection softening to
about the same level. This is especially important for using
adjustDetectorMap on quartzes, where the errors are typically
much smaller than this, which leads to overwhelming rejection.