convert ReferenceLineSet to pandas.DataFrame backend
It's just like ArcLineSet, so created a base class for general
purpose tables, and have ArcLineSet and ReferenceLineSet both
use it. TableBase uses a DataFrame under the covers, so it
hides all the complicated stuff in an API similar to the
old ArcLineSet/ReferenceLineSet, while providing access to
that complicated stuff if you want it. The main API changes
are that the constructor has changed (it takes a DataFrame,
which is what it's based on; the old constructor is called
fromRows), and the append method is gone (it would be very
inefficient). Additional functionality has been added, such
as slicing and indexing by numpy arrays.